[BBC-9084] Migrate `Stepper` to MDX
