+++ src/stepper/constants.ts
@@ -1,0 +1,9 @@
+export enum StepperDisplay {
+  SMALL = 'small',
+  LARGE = 'large',
+}
+
+export const StepperButtonSize = {
+  [StepperDisplay.SMALL]: 24,
+  [StepperDisplay.LARGE]: 48,
+}
